x64 compilation fix
git-svn-id: http://svn.miranda-ng.org/main/trunk@6545 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/WinterSpeak/src/EventInformation.h
@@ -39,7 +39,7 @@ class EventInformation
 	// Description : get the size of the message from the last event
 	// Return      : the size of the message
 	//--------------------------------------------------------------------------
-	unsigned int getMessageSize();
+	size_t getMessageSize();
 
 	//--------------------------------------------------------------------------
 	// Description : get the event string
